Removed: curl_global_cleanup(); from the normal exit after sucessful login detect step.
The reason of removing it is due to that it might cause the router stop from responding connections from clients to websites, which behaves very like a loss of DNS, but changing DNS on clients is not working (still not able to connect) (tested on OpenWRT 23.05.4 Mi Router 4A Gigabit Edition). It will happen after many rounds of running of this program, where most of running will quit the program normally after detecting network is reachable but not perform any login operations.